# [WIP] Optimize utf8 string length and byte offset functions