Replace existing utf8 strlen and byte offset functions with the UTF-8
length algorithm developed by George Pollard and discussed at

    https://www.daemonology.net/blog/2008-06-05-faster-utf8-strlen.html

The algorithm assumes that the String is valid UTF-8, which I believe is
enforced by `utf8_strlen`. The algorithm uses an ASCII fastpath and skips
characters based on their UTF-8 continuation bytes. This approach can be
more than twice as fast as the existing implementation.

The core algorithm is adapted for the following functions:

- utf8len
- mrb_utf8_len
- chars2bytes
- bytes2chars

This patch allows removing the utf8len_codepage static array.

Benchmarks (default build with MRB_UTF8_STRING defined)

`("aaa\n\nbbbbb\n\n\n\n\ncccc" * 100000000).reverse!`

master: 20.99s
patched: 7.36s

`("aaa\n\nbbbbb\n\n\n\n\ncccc" * 100000000)[1000000 - 1]`

master: 18.89s
patched: 8.14s

`("aaa\n\nbbbbb\n\n\n\n\ncccc" * 100000).each_line("\n") { }`

master: [does not complete in 4 minutes wall clock time]
patched: 0.02s

**NOTE**: I did not write the implementation of the algorithm, I only
adapted it to fit within the mruby APIs.

Fixes GH-4522.