@lopopolo 2019-06-23
I've generated a flamegraph with dtrace that compares master and this PR for the following benchmark:

```
./bin/mruby -e '100000.times { ("aaa\n\nbbbbb\n\n\n\n\ncccc" * 100000).each_line("\n") { } }'
```

## master

<img width="1378" alt="Screen Shot 2019-06-23 at 11 08 42 PM" src="https://user-images.githubusercontent.com/860434/59982553-f2926d80-960b-11e9-81a1-1ce4f3bb328e.png">

## PR 4529

<img width="1378" alt="Screen Shot 2019-06-23 at 11 09 01 PM" src="https://user-images.githubusercontent.com/860434/59982557-fe7e2f80-960b-11e9-8ea2-1464015d8ab6.png">