@shuujii 2019-06-24
It is interesting. However, it doesn't seem to work well in some cases:

#### Before patched with `MRB_UTF8_STRING` (same as Ruby)

    $ bin/mruby -e '["\xe3\x81", "a\x00z"].each{|s| p s.size}'
    2
    3

#### After patched with `MRB_UTF8_STRING`

    $ bin/mruby -e '["\xe3\x81", "a\x00z"].each{|s| p s.size}'
    1
    1