@dearblue 2019-06-24
@lopopolo

As a premise, I am imagined that policy of mruby is to handle binary data and strings without distinction.
Therefore, I think that even if there are mixed characters that are invalid as UTF-8, they need to be handled without problems.
I also want to hear other people's opinions about this.

----

Performing malicious tests this patch.

  - The character count of broken UTF-8 strings is different (This is similar to what @shuujii showed):

    ```console
    % ./build/host/bin/mruby -e 'p ("\xf1" * 5).size'
    2
    % ruby26 -e 'str = "\xf1" * 5; p str.encoding, str.size'
    #<Encoding:UTF-8>
    5
    ```

  - Crash when reversing a broken string as UTF-8:

    ```console
    % ./build/host/bin/mruby -e 'p ("\xf1" * 5).reverse'
    zsh: segmentation fault (core dumped)  ./mruby@test-utf8 -e 'p ("\xf1" * 5).reverse'
    ```

  - If missing trailing characters, it will make an out-of-range reference:

    ```console
    % ./build/host/bin/mruby -e 'p "\xf1"[0]'
    "\xf1\x00\x00\x00"
    ```