@dearblue 2019-06-24
`mrbtest` does not stop...

```console
% CFLAGS='-O0 -Wall -DMRB_UTF8_STRING' ./minirake -j10 clean `pwd`/build/test/bin/mrbtest
...snip...
% ./build/test/bin/mrbtest
mrbtest - Embeddable Ruby Test

.F..............................................................................F..F...X?..................F..F....F.......F.F.FFFF....F...^C
```