@lopopolo 2019-06-24
Oh no, it appears I was running the tests without `MRB_UTF8_STRING` defined. Thanks for the tip @dearblue. This patch is not ready for review. I'll mark it as WIP.