@lopopolo 2019-06-24
I've pushed my latest experiment. Not being able to store the encoding on the `RString` means we must process each character in the `char *` to determine the length. The existing approach with the lookup table is faster than what I've been able to come up with.

It looks like CRuby is able to fast-path UTF-8 strlen by storing whether the string has a "broken" encoding or not.

I'm going to abandon this PR. Thanks for taking a look.